moved css to another file and change unneccessary if statements
