+++ client/src/components/search/CourseSearchCSS.js
@@ -1,0 +1,42 @@
+const styles = {
+    filter: {
+        width: "400px",
+        display: "inline-block",
+        marginTop: 20,
+        marginLeft: 20,
+    },
+    button: {
+        display: "flex",
+        float: "center",
+        marginRight: 8,
+        marginTop: 8,
+        padding: "1px 8px 1px 8px",
+        fontSize: "11px",
+        boxShadow: "none",
+        borderRadius: "25px",
+        textTransform: "none",
+        width: "30%",
+        justifyContent: "space-around",
+    },
+    buttons: {
+        display: "flex",
+        marginLeft: 20,
+    },
+    searchBar: {
+        background: "#E4E8EE",
+        border: "2px solid #E4E8EE",
+        borderRadius: "15px 0px 0px 15px",
+        opacity: 1,
+        top: "100vw",
+        left: "20vw",
+        width: "28vw",
+        height: "43vw",
+    },
+    searchTxt: {
+        fontSize: "12px",
+        marginLeft: 20,
+        marginTop: 8,
+    },
+};
+
+export default styles;
